refactor: use rust workspace
