+++ rust/rust_c/Cargo.toml
@@ -5,66 +5,42 @@ edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
-[workspace]
-members = [
-    "src/kt_allocator",
-    "src/common",
-
-    "src/bitcoin",
-    "src/ethereum",
-    "src/solana",
-    "src/aptos",
-    "src/near",
-    "src/cosmos",
-    "src/cardano",
-    "src/sui",
-    "src/tron",
-    "src/xrp",
-    "src/arweave",
-    "src/ton",
-
-    "src/wallet",
-    "src/wallet/src/btc_only_wallet",
-    "src/wallet/src/multi_coins_wallet",
-
-    "src/simulator",
-
-    "src/test_cmd",
-    "src/test_cmd/src/btc_test_cmd",
-    "src/test_cmd/src/general_test_cmd",
-]
-
 [dependencies]
-kt_allocator = { path = "./src/kt_allocator" }
-
-common_rust_c = { path = "./src/common", default-features = false, optional = true }
-bitcoin_rust_c = { path = "./src/bitcoin", default-features = false, optional = true }
-wallet_rust_c = { path = "./src/wallet", default-features = false, optional = true }
+common_rust_c = { workspace = true, default-features = false, optional = true }
+bitcoin_rust_c = { workspace = true, default-features = false, optional = true }
+wallet_rust_c = { workspace = true, default-features = false, optional = true }
 test_cmd = { path = "./src/test_cmd", default-features = false, optional = true }
 
-ethereum_rust_c = { path = "./src/ethereum", optional = true }
-solana_rust_c = { path = "./src/solana", optional = true }
-aptos_rust_c = { path = "./src/aptos", optional = true }
-near_rust_c = { path = "./src/near", optional = true }
-cosmos_rust_c = { path = "./src/cosmos", optional = true }
-cardano_rust_c = { path = "./src/cardano", optional = true }
-sui_rust_c = { path = "./src/sui", optional = true }
-tron_rust_c = { path = "./src/tron", optional = true }
-xrp_rust_c = { path = "./src/xrp", optional = true }
-arweave_rust_c = { path = "./src/arweave", optional = true }
-stellar_rust_c = { path = "./src/stellar", optional = true }
-ton_rust_c = { path = "./src/ton", optional = true }
-simulator_rust_c = { path = "./src/simulator", optional = true, default-features = false }
+ethereum_rust_c = { workspace = true, optional = true }
+solana_rust_c = { workspace = true, optional = true }
+aptos_rust_c = { workspace = true, optional = true }
+near_rust_c = { workspace = true, optional = true }
+cosmos_rust_c = { workspace = true, optional = true }
+cardano_rust_c = { workspace = true, optional = true }
+sui_rust_c = { workspace = true, optional = true }
+tron_rust_c = { workspace = true, optional = true }
+xrp_rust_c = { workspace = true, optional = true }
+arweave_rust_c = { workspace = true, optional = true }
+stellar_rust_c = { workspace = true, optional = true }
+ton_rust_c = { workspace = true, optional = true }
+simulator_rust_c = { workspace = true, optional = true, default-features = false }
 
-[workspace.dependencies]
-sim_qr_reader = { path = "../sim_qr_reader"}
+cty = { workspace = true }
+cstr_core = { workspace = true }
 
 [lib]
 crate-type = ["staticlib"]
 
 [features]
+use-allocator = []
+firmware = ["use-allocator", "multi-coins"]
+firmware-btc-only = ["use-allocator", "btc-only"]
+simulator = ["multi-coins", "sui_rust_c/simulator", "simulator_rust_c/multi-coins"]
+simulator_btc_only = ["btc-only", "simulator_rust_c/btc-only"]
+default = ["simulator"]
+
 debug-memory = [
-    "default",
+    "multi-coins",
     "common_rust_c/debug-memory",
     "ethereum_rust_c/debug-memory",
     "bitcoin_rust_c/debug-memory",
@@ -80,10 +56,7 @@ debug-memory = [
     "xrp_rust_c/debug-memory",
     "ton_rust_c/debug-memory",
 ]
-simulator = ["multi-coins", "no-allocator", "sui_rust_c/simulator", "simulator_rust_c/multi-coins"]
-simulator_btc_only = ["btc-only", "no-allocator", "simulator_rust_c/btc-only"]
-no-allocator = []
-default = ["multi-coins"]
+
 multi-coins = [
     "common_rust_c/multi-coins",
     "bitcoin_rust_c/multi-coins",
