--- rust/rust_c/src/kt_allocator/.rustfmt.toml
@@ -1,3 +1,0 @@
-indent_style="Block"
-reorder_imports=true
-binop_separator="Front"
