--- rust/rust_c/src/kt_allocator/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "kt_allocator"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-cty = "0.2.2"
-cstr_core = "0.2.6"
